renamed and moved AreaViewPortController
